feat: split database files; add postgres
