+++ MatchZy.csproj
@@ -16,6 +16,7 @@
     <PackageReference Include="Dapper" Version="2.1.15" />
     <PackageReference Include="Microsoft.Data.Sqlite" Version="7.0.13" />
     <PackageReference Include="MySqlConnector" Version="2.3.0" />
+    <PackageReference Include="Npgsql" Version="8.0.5" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
     <PackageReference Include="SQLitePCLRaw.bundle_e_sqlite3" Version="2.1.6" />
   </ItemGroup>
